backend/requirements.txt
fastapi==0.63.0
uvicorn[standard]==0.13.4
pydantic==1.8.1

databases[postgresql]==0.4.2
tortoise-orm[asyncpg]==0.17.1
psycopg2>=2.8.6<2.9.0